requirements/base.txt
django==2.2.6
django-autocomplete-light
django-import-export==1.1.0
django-braces==1.13.0
django-pure-pagination==0.3.0
django-simple-captcha==0.5.10
django-debug-toolbar==1.11
boto3==1.9.156
django-storages==1.7.1
Pillow==5.2.0
django-cities-light==3.5.1
reportlab  
BeautifulSoup4
validators
pandas
django-cors-headers
django-htmlmin
